move non-webauthn code out of webauthn.go to separate files
